Break code into submodules
