+++ blues/__init__.py
@@ -6,4 +6,4 @@ BLUES
 # Define global version.
 from . import version
 __version__ = version.version
-from blues import moves, engine, ncmc_switching, simulation
+from blues import moves, engine, ncmc_switching, simulation, analysis
