+++ blues/analysis/1P8_1_0_c0.err
@@ -1,0 +1,7 @@
+### 1P8_1_0_c0 
+Traceback (most recent call last):
+  File "id-bindingmodes.py", line 89, in <module>
+    main(fpath,mol)
+  File "id-bindingmodes.py", line 67, in main
+    jsonfile = glob.glob('json/%s*.json' %molid)[0]
+IndexError: list index out of range
